small modifications on the BST incorporation front

+++ db_creation/reinitializeDB.R
@@ -2,11 +2,23 @@
 
 require(RPostgres)
 admin_db_param <- list(drv=Postgres(),dbname="postgres") 
-dbname <- "iavh_biol" 
+if("dbname" %in% ls())
+{
+  warning("database ", get("dbname")," will be reinitialized, if this is not what you want, please do something, fast!")
+  make_sleep<-T
+}else{
+  dbname <- "iavh_biol" 
+}
+if(make_sleep)
+{
+  Sys.sleep(60)
+}
 dbParam<-list(drv = Postgres(), dbname=dbname)
 # Admin connection: drop and create again the database
 dbAdmin<-do.call(dbConnect,args=admin_db_param)
-dbExecute(dbAdmin, paste0("DROP DATABASE ",dbQuoteIdentifier(dbAdmin,dbname)))
+if(dbname %in% dbGetQuery(dbAdmin,"SELECT datname FROM pg_database;")$datname){
+  dbExecute(dbAdmin, paste0("DROP DATABASE ",dbQuoteIdentifier(dbAdmin,dbname)))
+}
 dbExecute(dbAdmin, paste0("CREATE DATABASE ",dbQuoteIdentifier(dbAdmin,dbname)))
 dbDisconnect(dbAdmin)
 # Prepare database: postgis extension
